chore: move configs.rs to /src
