+++ mocks/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod cosmos;
 
+#[cfg(feature = "native")]
+pub mod configs;
 pub mod relayer;
 #[cfg(feature = "native")]
 pub mod sovereign;
