+++ mocks/src/utils/mod.rs
@@ -1,12 +1,7 @@
-#[cfg(feature = "native")]
-mod configs;
-
 mod mutex;
 
 use std::time::Duration;
 
-#[cfg(feature = "native")]
-pub use configs::*;
 pub use mutex::*;
 use tokio::time::sleep;
